requirements.txt
pip==25.1
streamlit==1.43.2
numpy==2.2.1
pandas==2.2.3
plotly==6.0.1
python-dotenv==1.0.1
snowflake-connector-python==3.12.4
pytz==2025.2
PyYAML==6.0.2
streamlit-authenticator==0.4.2
watchdog==6.0.0
Authlib==1.5.2
